refactor: rename .tsx files to UpperCamelCase to match their components
