+++ frontend/src/components/event/EventDetail.tsx
